Add files via upload
This project was done to demonstrate my capabilities to build a database using MySql and to use Symfony to build a online library. Very simple project for which i have created several controllers, forms as well as queries with the repositories. I have made a new release with a drop down menu in order to select only the relevant users from a particular library. I will upload the new video in order to explain this in more details.